pyproject.toml
[project]
name = "clean-dfine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mauro Sciancalepore", email = "maurosciancalepore98@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.1.0",
    "opencv-python-headless>=4.10.0.84",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "scipy>=1.13.1",
    "tensorboard>=2.18.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121", marker = "platform_system != 'Darwin'"},
]
torchvision = [
  { index = "pytorch-cu121", marker = "platform_system != 'Darwin'"},
]
